enlarge [min, max] range of GlQuantitative integer scale (use long long instead of int) + Histogram bug fix: ensure [min, max] range of double values can fit in [min, max] range of long long values before using an integer scale on an Histogram X axis
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@10073 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-ogl/include/tulip/GlQuantitativeAxis.h
@@ -63,7 +63,7 @@ public :
                          const LabelPosition &axisGradsLabelsPosition = LEFT_OR_BELOW, const bool drawFirstLabel = true);
 
 
-  void setAxisParameters(const int min, const int max, const unsigned int incrementStep,
+  void setAxisParameters(const long long min, const long long max, const unsigned long long incrementStep,
                          const LabelPosition &axisGradsLabelsPosition = LEFT_OR_BELOW, const bool drawFirstLabel = true);
 
 
@@ -138,7 +138,7 @@ private :
   bool logScale;
   unsigned int logBase;
   bool integerScale;
-  unsigned int incrementStep;
+  unsigned long long incrementStep;
   bool minMaxSet;
 
 };
